grltrackerplaylists: Fix removal of songs from a playlist
When a song is removed from a playlist, the song id is directly used
to remove the song from the playlist and update it. However, the song
in the playlist (the entry) is different from the song itself. Hence,
the song id cannot be used directly.

The issue is fixed by first retrieving the entry id and then use it to
remove the song from the playlist.