window: Fix album view selection by the stack switcher on launch
The first view added to the stack is the EmptyView which is not
visible on launch. Then, AlbumsView is added and becomes the
visible_child by default because it is visible. However, it is not
selected by the the stack switcher because the
"notify::visible-changed" signal is received before the child is added
to the stack switcher.

This issue is fixed by adding a "visible-child" notification. This
way, the stack switcher selected the AlbumView on launch.