gstplayer: Unschedule the clock when paused
When the player is paused, the clock "wait" times out and causes a
high cpu usage (this only happens with gstreamer 1.18).

This issue is fixed by unscheduling the clock id when going the player
is paused and then re-schedule it once the player resumes.

See https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/673

Related: #443