gstplayer: Unschedule the clock when stopped
This is a follow-up of commit ad31ba0f9c695ab8f42af9be252f789af0a39bdc
which unschedules the clock when the player is paused.

When the player is paused or stopped, the clock "wait" times out and
causes a high cpu usage (this only happens with gstreamer 1.18).

This commit unschedules the clock when the player is stopped.

See https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/673