playlistswidget: Use the row-activated signal to handle song clicks
SongWidget inherits from GtkEventBox and is inserted into a
GtkListBoxRow in order to be used by a GtkListBox. The SongWidget
consumers use the "button-release-event" signal from the GtkEventBox
to handle clicks on songs.

This commit changes the signal logic to use the "row-activated" signal
from the GtkListBox instead. It will make it possible to have
SongWidget directly inheriting from a GtkListBoxRow in the next
commit.