songwidget: Fix PlaylistContextMenu usage
With the recent SongWidget change to directly inherit from a
GtkListBoxRow, the rows became not selectable in order to ensure that
the play indicator stays visible when a row is activated. However,
this broke PlaylistContextMenu usage because it relies on rows being
selectable to work properly.

This issue is fixed by making SongWidget selectable and adding a css
rule to keep the play indicator visible when a SongWidget is
selected.