artstack: Set default size on stack
ArtStack currently sets the default size on a GtkImage in a stack page.
This works in GTK3. However, in GTK4 the hidden page is not used in size
calculation it seems, which results in incorrect sizing of some images.

Set the requested size on the ArtStack directly instead. Futureproofing
ArtStack sizing.