albumwidget: Do not include a ScrolledWindow
With the latest mockups, the AlbumWidget and the ArtistAlbumWidget
become very similar. Therefore, it becomes possible to use AlbumWidget
inside the ArtistsView. However, in the artists case, the AlbumWidget
cannot include a ScrolledWindow because the scrollbar from the
ArtistsView is already used.

This issue is fixed by removed the ScrolledWindow from the
AlbumWidget. A ScrolledWindow is inserted into the AlbumsView to
handle the album case.