artstack: Load default icon faster
Icon retrieval works by asynchronous lookup, the downside is that this
process is too unreliable for default icons to show up in time in the
views.

Work around this by setting a default icon for ArtStack in widgets right
away. For this add an art_type property, the property is needed to
differentiate between the different default icons for albums and
artists.

Related: #423