albumsview: Disable flowbox selection when selection is cancelled
When selection is cancelled the flowbox still has selection-mode as
Gtk.SelectionMode.MULTIPLE(but only if some album was selected through
its checkButton).  This results in the next click triggering selection
instead of opening the requested album.  The bug is not there when not
selecting by the checkButton since in that case
_on_selected_chidren_changed does set flowbox selection-mode to
Gtk.SelectionMode.NONE after processing the gesture.

Closes: #460