Make random shuffle actually random
Currently the randomness in the random shuffle is implemented by
returning a random result in a sort function, which is not very random
at all. In practice, it results in songs close to the start of the list
being sorted first.

To fix this, we can simply assign a random integer to each song and
compare them in the sorting function, to make the results consistent.

Fixes #369