grltrackerplaylists: Fix duplicated song removal
A playlist song that is duplicated gets removed with its duplicates.
For example, if you add an album twice and remove the first the other
song that is identical to it will also get removed, but this won't be
reflected in the view. To fix this we retrieve the song to be removed by
its position in the playlist rather than song_id since all songs that
are the same have the same song_id.

Closes: #469