grltrackerwrapper: Fix art callbacks
If an art uri is retrieved, the query callback will end with an empty
set (no media). Setting the thumbnail to generic here is unneeded as it
should be generic to begin with, but it also might overwrite a valid uri
that has been set in a previous query iteration.

Do not set the coreobject thumbnail to "generic".