albumwidget: Ensure that the correct album is played
When playing an album in artist view by clicking the play button, it
always starts at the first album. This is because no coresong is
requested: in that case, the first song of the first album is
automatically loaded.

This issue is fixed by loading the first song of the current album
when the coreobject is an artist.

Closes: https://gitlab.gnome.org/GNOME/gnome-music/-/issues/461