asyncqueue: Favour new queue additions
If an object got queued again, the old queue was retained. However, in
the case of ArtCache, this results in the new ArtCache queue being
discarded and the previous art being displayed. This lead to the wrong
art being displayed while cycling widgets in the listmodel views.

When requeing the same object, replace the old instance in the queue.

Fixes: #500