ui/SearchView: Hide albums and artists on init
While doing a quick initial search with no album and/or artist results,
the headers and flowbox would be visible and empty. This is because the
hiding is only triggered when their respective models update, which does
not happen with no results.

Make albums/artists invisible by default to not hit this problem.