artistart: Limit remote lookup rate
Music uses The Audio DB through Grilo for remote art lookup at this
time, which is for non-commercial use limited to 2 requests per second.
Also, Grilo is misbehaving with non-restricted lookups, with async
lookup calls never being returned.

Import the asynciolimiter source and use it to rate limit the lookups.