Clear search on empty string
When doing a repeat search of the same term after a view switch, the
results would not show up. This was because a similar search would not
change the results, not triggering search model updates and in turn not
notifying `SearchView` to show the results.

Always clear the search results on empty search, so a fresh search will
always trigger model updates.

Fixes: #610

Co-authored-by: Alireza Shabani <walt.shabani@gmail.com>