Drop Grilo and use LocalSearch directly
Since its inception Music has used Grilo to access metadata from various
sources. With the stagnation of the Grilo project and Music focusing on
just the LocalSearch (then Tracker) source for various reasons, it
became clear that the Grilo abstraction was becoming a burden.

This commit drops all use of Grilo, except for the art lookups.