Add ShuffleListModel for Queue
Implements a ShuffeListModel that can shuffle the items of the model
underneath. It is meant for the play queue in which the following items
from the played item are shuffled, but the previous items order remains
intact. When deshuffling at given point in the queue it will only order
the items not yet played in order as in the original playlist. This will
result in all items being played just once, no matter when shuffled or
deshuffled.

Also the activation paths in AlbumWidget were reworked:
* Clicking a SongWidget will play all songs in the view, so in artists
view it will play all songs of the artist. Clicking in a currently
playing album will continue the queue, otherwise it will be a new queue.
* Clicking the play button (or activating the menu option) will just
play the given album (even in artists view) as a new queue.

Fixes #572, #550