Add ShuffleListModel for Queue
Implements a ShuffeListModel that can shuffle the items of the model
underneath. It is meant for the play queue in which the following items
from the played item are shuffled, but the previous items order remains
intact. When deshuffling at given point in the queue it will only order
the items not yet played in order as in the original playlist. This will
result in all items being played just once, no matter when shuffled or
deshuffled.

Fixes #572