localsearch: Emit a change signal on every search result
If a search does not change the number of items found, SearchView will not
update the view. This resulted in the view not correctly updating its state.
For example, going directly to a search without hits, would keep the view
at its new search state as in both cases the number of items is 0.
This used to work before, but newer GTK probably tightened the signal emission.

Add an items-changed emission after every search.