grltrackerwrapper: Fix matching albums/artists with song title
For a long time search was slightly broken, it would not return an album
or artist that contains a song with the given search term. This used to
work, but seemingly got lost around the time of the core rewrite.

Add back song match for artist and album search queries.

Fixes #595